make sure no windows stuff gets compiled on linux
